Modularize Jquery-UI extension scripts and import into Common
